web_app/app/requirements.txt
pandas
streamlit
kafka-python
pydeck
plotly
sqlalchemy
pyhive
sasl
thrift
thrift-sasl